Address organization UI review feedback
