+++ src/locales/en.json
@@ -2,6 +2,8 @@
   "Child organizations": "Child organizations",
   "Create": "Create",
   "Create organization": "Create organization",
+  "Create your first organization": "Create your first organization",
+  "Create your first organization to start building your internal company hierarchy.": "Create your first organization to start building your internal company hierarchy.",
   "External system ID": "External system ID",
   "Failed to create organization.": "Failed to create organization.",
   "Failed to rename organization.": "Failed to rename organization.",
@@ -12,6 +14,7 @@
   "No child organizations": "No child organizations",
   "No facilities are owned by this organization": "No facilities are owned by this organization",
   "No organizations found": "No organizations found",
+  "No organizations yet": "No organizations yet",
   "No parent (root organization)": "No parent (root organization)",
   "Organization details": "Organization details",
   "Organization has a relationship to a missing child: {partyId}": "Organization has a relationship to a missing child: {partyId}",
