+++ src/views/Organizations.vue
@@ -33,7 +33,7 @@
             detail
           >
             <ion-label class="ion-text-wrap">
-              <h2>{{ node.groupName || node.partyId }}</h2>
+              {{ node.groupName || node.partyId }}
               <p>{{ node.path.join(" / ") }}</p>
             </ion-label>
             <ion-badge v-if="node.partyId === primaryOrganizationId" slot="end">
@@ -58,42 +58,117 @@
           </ion-card-content>
         </ion-card>
 
-        <ion-list>
-          <ul class="organization-tree" aria-label="Organization hierarchy">
-            <OrganizationTreeItem
-              v-for="root in forest.roots"
-              :key="root.partyId"
-              :node="root"
-              :primary-id="primaryOrganizationId"
-            />
-          </ul>
+        <ion-list :aria-label="translate('Organization hierarchy')">
+          <OrganizationTreeItem
+            v-for="root in forest.roots"
+            :key="root.partyId"
+            :node="root"
+            :primary-id="primaryOrganizationId"
+          />
         </ion-list>
       </main>
 
-      <main v-else class="empty-state">
-        {{ translate("No organizations found") }}
+      <main v-else>
+        <ion-card class="ion-text-center">
+          <ion-card-header>
+            <ion-icon :icon="businessOutline" color="medium" size="large" />
+            <ion-card-title>{{ translate("No organizations yet") }}</ion-card-title>
+          </ion-card-header>
+          <ion-card-content>
+            <p>
+              {{ translate("Create your first organization to start building your internal company hierarchy.") }}
+            </p>
+            <ion-button v-if="canManage" class="ion-margin-top" @click="openCreateModal()">
+              {{ translate("Create your first organization") }}
+              <ion-icon slot="end" :icon="addOutline" />
+            </ion-button>
+          </ion-card-content>
+        </ion-card>
       </main>
 
       <ion-fab v-if="canManage" slot="fixed" vertical="bottom" horizontal="end">
-        <ion-fab-button :aria-label="translate('Create organization')" @click="showCreate = true">
+        <ion-fab-button :aria-label="translate('Create organization')" @click="openCreateModal()">
           <ion-icon :icon="addOutline" />
         </ion-fab-button>
       </ion-fab>
 
-      <CreateOrganizationModal
-        :is-open="showCreate"
-        :organizations="organizations"
-        @dismiss="showCreate = false"
-        @created="organizationCreated"
-      />
+      <ion-modal :is-open="showCreate" @did-dismiss="closeCreateModal()">
+        <ion-header>
+          <ion-toolbar>
+            <ion-buttons slot="start">
+              <ion-button @click="closeCreateModal()">
+                {{ translate("Cancel") }}
+              </ion-button>
+            </ion-buttons>
+            <ion-title>{{ translate("Create organization") }}</ion-title>
+            <ion-buttons slot="end">
+              <ion-button :disabled="saving || !canCreate" @click="saveOrganization()">
+                {{ translate("Create") }}
+              </ion-button>
+            </ion-buttons>
+          </ion-toolbar>
+        </ion-header>
+        <ion-content>
+          <ion-list>
+            <ion-item>
+              <ion-input
+                v-model="groupName"
+                :label="translate('Organization name')"
+                label-placement="stacked"
+                :maxlength="100"
+                required
+              />
+            </ion-item>
+            <ion-item>
+              <ion-input
+                v-model="partyId"
+                :label="translate('Organization ID')"
+                :helper-text="translate('Generated from the name. You can edit it before creating the organization.')"
+                label-placement="stacked"
+                :maxlength="20"
+                required
+                @ion-input="partyIdManuallyEdited = true"
+              />
+            </ion-item>
+            <ion-item>
+              <ion-input
+                v-model="externalId"
+                :label="translate('External system ID')"
+                label-placement="stacked"
+              />
+            </ion-item>
+            <ion-item>
+              <ion-select
+                v-model="parentPartyId"
+                :label="translate('Parent organization')"
+                label-placement="stacked"
+                interface="popover"
+              >
+                <ion-select-option value="">
+                  {{ translate("No parent (root organization)") }}
+                </ion-select-option>
+                <ion-select-option
+                  v-for="organization in organizations"
+                  :key="organization.partyId"
+                  :value="organization.partyId"
+                >
+                  {{ organization.groupName || organization.partyId }}
+                </ion-select-option>
+              </ion-select>
+            </ion-item>
+          </ion-list>
+        </ion-content>
+      </ion-modal>
     </ion-content>
   </ion-page>
 </template>
 
 <script setup lang="ts">
-import { translate } from "@common";
+import { commonUtil, translate } from "@common";
 import {
   IonBadge,
+  IonButton,
+  IonButtons,
   IonCard,
   IonCardContent,
   IonCardHeader,
@@ -103,29 +178,35 @@ import {
   IonFabButton,
   IonHeader,
   IonIcon,
+  IonInput,
   IonItem,
   IonLabel,
   IonList,
   IonListHeader,
   IonMenuButton,
+  IonModal,
   IonPage,
   IonSearchbar,
+  IonSelect,
+  IonSelectOption,
   IonSkeletonText,
   IonTitle,
   IonToolbar,
   onIonViewWillEnter,
 } from "@ionic/vue";
-import { addOutline } from "ionicons/icons";
-import { computed, ref } from "vue";
-import CreateOrganizationModal from "@/components/organization/CreateOrganizationModal.vue";
+import { addOutline, businessOutline } from "ionicons/icons";
+import { computed, ref, watch } from "vue";
 import OrganizationTreeItem from "@/components/organization/OrganizationTreeItem.vue";
 import {
   type OrganizationAnomaly,
+  createOrganization,
+  suggestOrganizationId,
   useOrganizations,
   usePrimaryOrganization,
 } from "@/composables/useOrganizations";
 import { useAuth } from "@/composables/useSecurity";
 import router from "@/router";
+import { getResponseErrorMessage } from "@/utils";
 
 const { organizations, forest, hydrated } = useOrganizations();
 const { primaryOrganizationId, load: loadPrimaryOrganization } = usePrimaryOrganization();
@@ -133,6 +214,13 @@ const { hasPermission } = useAuth();
 const canManage = computed(() => hasPermission("PARTYMGR_ADMIN"));
 const searchText = ref("");
 const showCreate = ref(false);
+const groupName = ref("");
+const partyId = ref("");
+const externalId = ref("");
+const parentPartyId = ref("");
+const partyIdManuallyEdited = ref(false);
+const saving = ref(false);
+const canCreate = computed(() => Boolean(groupName.value.trim() && partyId.value.trim()));
 
 const searchResults = computed(() => {
   const term = searchText.value.trim().toLowerCase();
@@ -165,9 +253,39 @@ function anomalyMessage(anomaly: OrganizationAnomaly): string {
   return translate("Organization hierarchy contains a cycle at: {partyId}", { partyId: anomaly.partyId });
 }
 
-function organizationCreated(partyId: string) {
+watch(groupName, (name) => {
+  if(!partyIdManuallyEdited.value) {partyId.value = suggestOrganizationId(name);}
+});
+
+function openCreateModal() {
+  groupName.value = "";
+  partyId.value = "";
+  externalId.value = "";
+  parentPartyId.value = "";
+  partyIdManuallyEdited.value = false;
+  showCreate.value = true;
+}
+
+function closeCreateModal() {
   showCreate.value = false;
-  void router.push(`/organization-details/${encodeURIComponent(partyId)}`);
+}
+
+async function saveOrganization() {
+  saving.value = true;
+  try {
+    const createdPartyId = await createOrganization({
+      partyId: partyId.value,
+      groupName: groupName.value,
+      externalId: externalId.value,
+      parentPartyId: parentPartyId.value || undefined,
+    });
+    closeCreateModal();
+    await router.push(`/organization-details/${encodeURIComponent(createdPartyId)}`);
+  } catch (error) {
+    await commonUtil.showToast(getResponseErrorMessage(error, translate("Failed to create organization.")));
+  } finally {
+    saving.value = false;
+  }
 }
 
 onIonViewWillEnter(() => void loadPrimaryOrganization());
@@ -178,13 +296,4 @@ ion-content {
   --padding-bottom: var(--spacer-2xl);
 }
 
-.organization-tree {
-  margin: 0;
-  padding: 0;
-}
-
-.empty-state {
-  padding: var(--spacer-base);
-  text-align: center;
-}
 </style>
